refactor: consolidate date helpers; address review on quick-add VM
Review follow-ups (#474):
- quick-add-view-model: drop the obsolete pushup-sentinel special-casing
  (pushup is a first-class catalog exercise now, so findExerciseDefinition
  /exerciseDisplayName resolve it to the same label + fitness_center icon)
  and fix the misleading `label` doc (auto-count uses the bare exercise
  label, no "Auto:" prefix).

Consolidation:
- Move the duplicated daysBetween + sortedUniqueDates into @pu-stats/date
  (with specs); analysis/trend-math and dashboard/dashboard-math now
  import the shared helpers instead of each keeping a private copy.

Behaviour-preserving; lib + web tests pass.

Refs #456
